BJS2-66141: с учетом некоторых комментариев
